Change the namespace to siddhi.execution.extrema
