+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.wso2.extension.siddhi</groupId>
-    <artifactId>org.wso2.extesion.siddhi.window.minbymaxby</artifactId>
+    <artifactId>org.wso2.extesion.siddhi.window.extrema</artifactId>
     <version>1.0.0-SNAPSHOT</version>
