+++ src/main/java/org/wso2/extension/siddhi/execution/extrema/MaxByMinByLengthWindowProcessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -14,11 +14,13 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
+ *
  */
-package org.wso2.extension.siddhi.window.minbymaxby.lengthwindow.length;
 
-import org.wso2.extension.siddhi.window.minbymaxby.MaxByMinByConstants;
-import org.wso2.extension.siddhi.window.minbymaxby.MaxByMinByExecutor;
+package org.wso2.extension.siddhi.execution.extrema;
+
+import org.wso2.extension.siddhi.execution.extrema.util.MaxByMinByConstants;
+import org.wso2.extension.siddhi.execution.extrema.util.MaxByMinByExecutor;
 import org.wso2.siddhi.core.config.ExecutionPlanContext;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
 import org.wso2.siddhi.core.event.state.StateEvent;
