+++ src/main/java/org/wso2/extension/siddhi/execution/extrema/util/MaxByMinByConstants.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.extension.siddhi.window.minbymaxby;
+package org.wso2.extension.siddhi.execution.extrema.util;
 
 /**
  * Holds the constants required to retrieve minimum or maximum event
