+++ src/main/resources/extrema.siddhiext
@@ -1,0 +1,26 @@
+#
+# Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+minByLengthBatch=org.wso2.extension.siddhi.execution.extrema.MinByLengthBatchWindowProcessor
+maxByLengthBatch=org.wso2.extension.siddhi.execution.extrema.MaxByLengthBatchWindowProcessor
+minByLength=org.wso2.extension.siddhi.execution.extrema.MinByLengthWindowProcessor
+maxByLength=org.wso2.extension.siddhi.execution.extrema.MaxByLengthWindowProcessor
+minbytimebatch=org.wso2.extension.siddhi.execution.extrema.MinByTimeBatchWindowProcessor
+maxbytimebatch=org.wso2.extension.siddhi.execution.extrema.MaxByTimeBatchWindowProcessor
+minbytime=org.wso2.extension.siddhi.execution.extrema.MinByTimeWindowProcessor
+maxbytime=org.wso2.extension.siddhi.execution.extrema.MaxByTimeWindowProcessor
