--- src/main/resources/minbymaxby.siddhiext
@@ -1,26 +1,0 @@
-#
-# Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-#
-# WSO2 Inc. licenses this file to you under the Apache License,
-# Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-minByLengthBatch=org.wso2.extension.siddhi.window.minbymaxby.lengthwindow.lengthbatch.MinByLengthBatchWindowProcessor
-maxByLengthBatch=org.wso2.extension.siddhi.window.minbymaxby.lengthwindow.lengthbatch.MaxByLengthBatchWindowProcessor
-minByLength=org.wso2.extension.siddhi.window.minbymaxby.lengthwindow.length.MinByLengthWindowProcessor
-maxByLength=org.wso2.extension.siddhi.window.minbymaxby.lengthwindow.length.MaxByLengthWindowProcessor
-minbytimebatch=org.wso2.extension.siddhi.window.minbymaxby.TimeWindow.MinByTimeBatchWindowProcessor
-maxbytimebatch=org.wso2.extension.siddhi.window.minbymaxby.TimeWindow.MaxByTimeBatchWindowProcessor
-minbytime=org.wso2.extension.siddhi.window.minbymaxby.TimeWindow.MinByTimeWindowProcessor
-maxbytime=org.wso2.extension.siddhi.window.minbymaxby.TimeWindow.MaxByTimeWindowProcessor
