+++ src/test/java/org/wso2/extension/siddhi/execution/extrema/MaxByLengthBatchWindowProcessorTestCase.java
@@ -1,6 +1,6 @@
 
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,8 +15,10 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
+ *
  */
-package org.wso2.extension.siddhi.window.minbymaxby.lengthwindow;
+
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.apache.log4j.Logger;
 import org.junit.Before;
@@ -24,7 +26,6 @@ import org.junit.Test;
 import org.wso2.siddhi.core.ExecutionPlanRuntime;
 import org.wso2.siddhi.core.SiddhiManager;
 import org.wso2.siddhi.core.event.Event;
-import org.wso2.siddhi.core.query.output.callback.QueryCallback;
 import org.wso2.siddhi.core.stream.input.InputHandler;
 import org.wso2.siddhi.core.stream.output.StreamCallback;
 import org.wso2.siddhi.core.util.EventPrinter;
@@ -35,12 +36,11 @@ import java.util.List;
 import static org.junit.Assert.assertArrayEquals;
 
 /**
- * Created by mathuriga on 01/10/16.
+ * Created by mathuriga on 28/09/16.
  */
-public class MaxByLengthWindowProcessorTestCase {
-    private static final Logger log = Logger.getLogger(MaxByLengthWindowProcessorTestCase.class);
-    int count;
-    List<Object> results = new ArrayList<Object>();
+public class MaxByLengthBatchWindowProcessorTestCase {
+    private static final Logger log = Logger.getLogger(MaxByLengthBatchWindowProcessorTestCase.class);
+    private int count;
 
     @Before
     public void init() {
@@ -49,18 +49,14 @@ public class MaxByLengthWindowProcessorTestCase {
 
 
     @Test
-    public void testMaxByLengthWindowProcessor1() throws InterruptedException {
-        log.info("Testing maxByLengthWindowProcessor with no of events less than window size for float type parameter");
+    public void testMaxByWindowForLengthBatch() throws InterruptedException {
+        log.info("Testing maxByLengthBatchWindowProcessor with no of events equal to window size for float parameter");
+
         SiddhiManager siddhiManager = new SiddhiManager();
-
-
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLength(price, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLengthBatch(price, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
-        results.add(new Object[]{"IBM", 700f, 14});
-        results.add(new Object[]{"IBM", 700f, 14});
-        results.add(new Object[]{"WSO2", 790f, 1});
         try {
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
 
@@ -68,83 +64,44 @@ public class MaxByLengthWindowProcessorTestCase {
                 public void receive(Event[] events) {
                     System.out.print("output event: ");
                     EventPrinter.print(events);
+                    Object[] results = new Object[]{"dg", 900f, 24};
+                    assertArrayEquals(results, events[0].getData());
 
-                    for (Event event : events) {
-                        assertArrayEquals((Object[]) results.get(count), event.getData());
-                        count++;
-                    }
                 }
             });
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
             inputHandler.send(new Object[]{"IBM", 700f, 14});
-            inputHandler.send(new Object[]{"IBM", 20.5f, 2});
-            inputHandler.send(new Object[]{"WSO2", 790f, 1});
+            inputHandler.send(new Object[]{"IBM", 40.5f, 2});
+            inputHandler.send(new Object[]{"et", 900f, 1});
+            inputHandler.send(new Object[]{"dg", 900f, 24});
+
             Thread.sleep(1000);
+
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
     @Test
-    public void testMaxByLengthWindowProcessor10() throws InterruptedException {
-        log.info("Testing maxByLengthWindowProcessor with no of events less than window size for float type parameter");
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-
-        String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLength(price, 2) select symbol,price," +
-                "volume insert into outputStream ;";
-        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
-//        results.add(new Object[]{"IBM", 700f, 14});
-//        results.add(new Object[]{"IBM", 700f, 14});
-//        results.add(new Object[]{"WSO2", 790f, 1});
-        try {
-            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
-
-                @Override
-                public void receive(Event[] events) {
-                    System.out.print("output event: ");
-                    EventPrinter.print(events);
-
-                    for (Event event : events) {
-                        //  assertArrayEquals((Object[]) results.get(count), event.getData());
-                        count++;
-                    }
-                }
-            });
-            InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
-            executionPlanRuntime.start();
-            inputHandler.send(new Object[]{"IBM", 10f, 14});
-            inputHandler.send(new Object[]{"IBM", 20.5f, 2});
-            inputHandler.send(new Object[]{"WSO2", 20.5f, 1});
-            inputHandler.send(new Object[]{"WSO2", 23f, 1});
-
-            Thread.sleep(1000);
-        } finally {
-            executionPlanRuntime.shutdown();
-        }
-    }
-
-
-    @Test
-    public void testMaxByLengthWindowProcessor2() throws InterruptedException {
-        log.info("Testing maxByLengthWindowProcessor with no of events equal to window size for integer type parameter");
+    public void testMaxByWindowForLengthBatch2() throws InterruptedException {
+        log.info("Testing maxByLengthBatchWindowProcessor with no of events greater than window size for int parameter");
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLength(volume, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLengthBatch(volume, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         try {
             final List<Object> results = new ArrayList<Object>();
-            results.add(new Object[]{"IBM", 700f, 14});
-            results.add(new Object[]{"IBM", 700f, 14});
-            results.add(new Object[]{"IBM", 700f, 20});
-            results.add(new Object[]{"ZZZ", 60.5f, 82});
+            results.add(new Object[]{"IBM", 700f, 142});
+            results.add(new Object[]{"dg", 60.5f, 24});
+
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+
                 @Override
                 public void receive(Event[] events) {
+
                     System.out.print("output event: ");
                     EventPrinter.print(events);
                     for (Event event : events) {
@@ -156,71 +113,72 @@ public class MaxByLengthWindowProcessorTestCase {
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
             inputHandler.send(new Object[]{"IBM", 700f, 14});
-            inputHandler.send(new Object[]{"IBM", 60.5f, 12});
-            inputHandler.send(new Object[]{"IBM", 700f, 20});
-            inputHandler.send(new Object[]{"ZZZ", 60.5f, 82});
+            inputHandler.send(new Object[]{"IBM", 60.5f, 2});
+            inputHandler.send(new Object[]{"IBM", 700f, 142});
+            inputHandler.send(new Object[]{"IBM", 60.5f, 21});
+            inputHandler.send(new Object[]{"et", 700f, 1});
+            inputHandler.send(new Object[]{"dg", 60.5f, 24});
+            inputHandler.send(new Object[]{"IBM", 60.5f, 21});
+            inputHandler.send(new Object[]{"et", 700f, 1});
+            inputHandler.send(new Object[]{"dg", 60.5f, 24});
             Thread.sleep(1000);
+
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
-
     @Test
-    public void testMaxByLengthWindowProcessor3() throws InterruptedException {
-        log.info("Testing maxByLengthWindowProcessor with no of events greater than window size for String type parameter");
+    public void testMaxByWindowForLengthBatch3() throws InterruptedException {
+        log.info("Testing maxByLengthBatchWindowProcessor with no of events greater than window size for String parameter");
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLength(symbol, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLengthBatch(symbol, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         try {
             final List<Object> results = new ArrayList<Object>();
-            results.add(new Object[]{"bbc", 700f, 14});
-            results.add(new Object[]{"bbc", 700f, 14});
-            results.add(new Object[]{"xxx", 700f, 2});
-            results.add(new Object[]{"xxx", 700f, 2});
-            results.add(new Object[]{"xxx", 700f, 2});
-            results.add(new Object[]{"zzz", 60.5f, 12});
-            results.add(new Object[]{"zzz", 60.5f, 12});
-            results.add(new Object[]{"zzz", 60.5f, 12});
-            results.add(new Object[]{"zzz", 60.5f, 12});
-            results.add(new Object[]{"rye", 60.5f, 82});
+            results.add(new Object[]{"XXX", 700f, 14});
+            results.add(new Object[]{"YTX", 60.5f, 24});
 
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+
                 @Override
                 public void receive(Event[] events) {
+
                     System.out.print("output event: ");
                     EventPrinter.print(events);
-
                     for (Event event : events) {
                         assertArrayEquals((Object[]) results.get(count), event.getData());
                         count++;
                     }
+
                 }
             });
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
-            inputHandler.send(new Object[]{"bbc", 700f, 14});
-            inputHandler.send(new Object[]{"bab", 60.5f, 12});
-            inputHandler.send(new Object[]{"xxx", 700f, 2});
-            inputHandler.send(new Object[]{"ddd", 60.5f, 82});
-            inputHandler.send(new Object[]{"abc", 700f, 84});
-            inputHandler.send(new Object[]{"zzz", 60.5f, 12});
-            inputHandler.send(new Object[]{"aaa", 700f, 2});
-            inputHandler.send(new Object[]{"dhh", 60.5f, 82});
-            inputHandler.send(new Object[]{"drg", 700f, 2});
-            inputHandler.send(new Object[]{"rye", 60.5f, 82});
+            inputHandler.send(new Object[]{"XXX", 700f, 14});
+            inputHandler.send(new Object[]{"ABC", 60.5f, 2});
+            inputHandler.send(new Object[]{"AAA", 700f, 142});
+            inputHandler.send(new Object[]{"ACD", 60.5f, 21});
+            System.out.println("-------------------");
+            inputHandler.send(new Object[]{"RTE", 700f, 1});
+            inputHandler.send(new Object[]{"YTX", 60.5f, 24});
+            inputHandler.send(new Object[]{"DGF", 60.5f, 21});
+            inputHandler.send(new Object[]{"ETR", 700f, 1});
+            inputHandler.send(new Object[]{"DXD", 60.5f, 24});
+
             Thread.sleep(1000);
+
+
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
-
     @Test
-    public void testMaxByWindowForLength4() throws InterruptedException {
+    public void testMaxByWindowForLengthBatch4() throws InterruptedException {
         log.info("Join test1");
 
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -229,10 +187,11 @@ public class MaxByLengthWindowProcessorTestCase {
                 "define stream twitterStream (user string, tweet string, company string,likes int); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.lengthBatch(3) join twitterStream#window.minbymaxby:maxByLength(likes, 2) " +
+                "from cseEventStream#window.lengthBatch(3) join twitterStream#window.extrema:maxByLengthBatch(likes, 2) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price, twitterStream.likes " +
                 "insert into outputStream ;";
+
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
         try {
             final List<Object> results = new ArrayList<Object>();
@@ -249,12 +208,12 @@ public class MaxByLengthWindowProcessorTestCase {
             InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
             executionPlanRuntime.start();
 
-            cseEventStreamHandler.send(new Object[]{"XXX", 699f, 14});
+            cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
             cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
             cseEventStreamHandler.send(new Object[]{"WSO2", 700f, 142});
 
-            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2", 43});
-            twitterStreamHandler.send(new Object[]{"User1", "Hello SIDDHI", "WSO2", 45});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2", 23});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello SIDDHI", "WSO2", 56});
 
             cseEventStreamHandler.send(new Object[]{"ACD", 60.5f, 21});
             cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
@@ -266,7 +225,7 @@ public class MaxByLengthWindowProcessorTestCase {
     }
 
     @Test
-    public void testMaxByWindowForLength6() throws InterruptedException {
+    public void testMaxByWindowForLengthBatch6() throws InterruptedException {
         log.info("Join test2");
 
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -275,10 +234,10 @@ public class MaxByLengthWindowProcessorTestCase {
                 "define stream twitterStream (num float, tweet string, company string); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.minbymaxby:maxByLength(price, 2) join twitterStream#window.minbymaxby:maxByLength(num, 2) " +
+                "from cseEventStream#window.extrema:maxByLengthBatch(price, 2) join twitterStream#window.extrema:maxByLengthBatch(num, 2) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
-                "insert all events into outputStream ;";
+                "insert events into outputStream ;";
 
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
         try {
@@ -289,54 +248,7 @@ public class MaxByLengthWindowProcessorTestCase {
                 public void receive(Event[] events) {
 
                     System.out.print("output event: ");
-//                    if(events==null){
-//                        System.out.println("There is no output events");
-//                    }
                     EventPrinter.print(events);
-
-                }
-            });
-            InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
-            InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
-            executionPlanRuntime.start();
-
-            cseEventStreamHandler.send(new Object[]{"WSO2", 10f, 14});
-            cseEventStreamHandler.send(new Object[]{"AAA", 60.5f, 2});
-
-            twitterStreamHandler.send(new Object[]{100, "Hello World", "XXX"});
-            twitterStreamHandler.send(new Object[]{101, "Hello SIDDHI", "WSO2"});
-            twitterStreamHandler.send(new Object[]{104, "Hello CEP", "WSO2"});
-
-
-        } finally {
-            executionPlanRuntime.shutdown();
-        }
-
-    }
-
-
-    @Test
-    public void testMaxByWindowForLength7() throws InterruptedException {
-        log.info("Join test3");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String streams = "" +
-                "define stream cseEventStream (symbol string, price float, volume int); " +
-                "define stream twitterStream (num float, tweet string, company string); ";
-        String query = "" +
-                "@info(name = 'query1') " +
-                "from cseEventStream#window.minbymaxby:maxByLength(price, 2) join twitterStream#window.Length(2) " +
-                "on cseEventStream.symbol== twitterStream.company " +
-                "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
-                "insert events into outputStream ;";
-
-        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
-        try {
-
-            executionPlanRuntime.addCallback("query1", new QueryCallback() {
-                @Override
-                public void receive(long l, Event[] events, Event[] events1) {
-                    EventPrinter.print(l, events, events1);
                 }
             });
             InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
@@ -348,13 +260,10 @@ public class MaxByLengthWindowProcessorTestCase {
 
             twitterStreamHandler.send(new Object[]{100, "Hello World", "XXX"});
             twitterStreamHandler.send(new Object[]{101, "Hello SIDDHI", "WSO2"});
-            Thread.sleep(100);
-            System.out.println(",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,");
 
-            cseEventStreamHandler.send(new Object[]{"WSO2", 900f, 14});
-            System.out.println("------------------------------------");
-            cseEventStreamHandler.send(new Object[]{"XXX", 70.5f, 2});
-
+            cseEventStreamHandler.send(new Object[]{"ACD", 60.5f, 21});
+            cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
+            cseEventStreamHandler.send(new Object[]{"WSO2", 60.5f, 222});
 
         } finally {
             executionPlanRuntime.shutdown();
@@ -363,4 +272,49 @@ public class MaxByLengthWindowProcessorTestCase {
     }
 
 
+    @Test
+    public void testMaxByWindowForLengthBatch7() throws InterruptedException {
+        log.info("Join test3");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String streams = "" +
+                "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define stream twitterStream (num float, tweet string, company string); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.extrema:maxByLengthBatch(price, 2) join twitterStream#window.LengthBatch(2) " +
+                "on cseEventStream.symbol== twitterStream.company " +
+                "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
+                "insert events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
+        try {
+            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+
+                @Override
+                public void receive(Event[] events) {
+
+                    System.out.print("output event: ");
+                    EventPrinter.print(events);
+
+                }
+            });
+            InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+            InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
+            executionPlanRuntime.start();
+
+            cseEventStreamHandler.send(new Object[]{"WSO2", 700f, 14});
+            cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
+
+            twitterStreamHandler.send(new Object[]{100, "Hello World", "XXX"});
+            twitterStreamHandler.send(new Object[]{101, "Hello SIDDHI", "WSO2"});
+
+            cseEventStreamHandler.send(new Object[]{"WSO2", 900f, 14});
+            cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
+
+        } finally {
+            executionPlanRuntime.shutdown();
+        }
+
+    }
 }
