+++ src/test/java/org/wso2/extension/siddhi/execution/extrema/MaxByLengthWindowProcessorTestCase.java
@@ -1,6 +1,6 @@
 
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,8 +15,10 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
+ *
  */
-package org.wso2.extension.siddhi.window.minbymaxby.lengthwindow;
+
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.apache.log4j.Logger;
 import org.junit.Before;
@@ -24,6 +26,7 @@ import org.junit.Test;
 import org.wso2.siddhi.core.ExecutionPlanRuntime;
 import org.wso2.siddhi.core.SiddhiManager;
 import org.wso2.siddhi.core.event.Event;
+import org.wso2.siddhi.core.query.output.callback.QueryCallback;
 import org.wso2.siddhi.core.stream.input.InputHandler;
 import org.wso2.siddhi.core.stream.output.StreamCallback;
 import org.wso2.siddhi.core.util.EventPrinter;
@@ -34,11 +37,12 @@ import java.util.List;
 import static org.junit.Assert.assertArrayEquals;
 
 /**
- * Created by mathuriga on 28/09/16.
+ * Created by mathuriga on 01/10/16.
  */
-public class MaxByLengthBatchWindowProcessorTestCase {
-    private static final Logger log = Logger.getLogger(MaxByLengthBatchWindowProcessorTestCase.class);
-    private int count;
+public class MaxByLengthWindowProcessorTestCase {
+    private static final Logger log = Logger.getLogger(MaxByLengthWindowProcessorTestCase.class);
+    int count;
+    List<Object> results = new ArrayList<Object>();
 
     @Before
     public void init() {
@@ -47,14 +51,18 @@ public class MaxByLengthBatchWindowProcessorTestCase {
 
 
     @Test
-    public void testMaxByWindowForLengthBatch() throws InterruptedException {
-        log.info("Testing maxByLengthBatchWindowProcessor with no of events equal to window size for float parameter");
-
+    public void testMaxByLengthWindowProcessor1() throws InterruptedException {
+        log.info("Testing maxByLengthWindowProcessor with no of events less than window size for float type parameter");
         SiddhiManager siddhiManager = new SiddhiManager();
+
+
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLengthBatch(price, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLength(price, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+        results.add(new Object[]{"IBM", 700f, 14});
+        results.add(new Object[]{"IBM", 700f, 14});
+        results.add(new Object[]{"WSO2", 790f, 1});
         try {
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
 
@@ -62,44 +70,83 @@ public class MaxByLengthBatchWindowProcessorTestCase {
                 public void receive(Event[] events) {
                     System.out.print("output event: ");
                     EventPrinter.print(events);
-                    Object[] results = new Object[]{"dg", 900f, 24};
-                    assertArrayEquals(results, events[0].getData());
 
+                    for (Event event : events) {
+                        assertArrayEquals((Object[]) results.get(count), event.getData());
+                        count++;
+                    }
                 }
             });
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
             inputHandler.send(new Object[]{"IBM", 700f, 14});
-            inputHandler.send(new Object[]{"IBM", 40.5f, 2});
-            inputHandler.send(new Object[]{"et", 900f, 1});
-            inputHandler.send(new Object[]{"dg", 900f, 24});
-
+            inputHandler.send(new Object[]{"IBM", 20.5f, 2});
+            inputHandler.send(new Object[]{"WSO2", 790f, 1});
             Thread.sleep(1000);
-
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
     @Test
-    public void testMaxByWindowForLengthBatch2() throws InterruptedException {
-        log.info("Testing maxByLengthBatchWindowProcessor with no of events greater than window size for int parameter");
-
+    public void testMaxByLengthWindowProcessor10() throws InterruptedException {
+        log.info("Testing maxByLengthWindowProcessor with no of events less than window size for float type parameter");
         SiddhiManager siddhiManager = new SiddhiManager();
+
+
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLengthBatch(volume, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLength(price, 2) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+//        results.add(new Object[]{"IBM", 700f, 14});
+//        results.add(new Object[]{"IBM", 700f, 14});
+//        results.add(new Object[]{"WSO2", 790f, 1});
         try {
-            final List<Object> results = new ArrayList<Object>();
-            results.add(new Object[]{"IBM", 700f, 142});
-            results.add(new Object[]{"dg", 60.5f, 24});
-
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
 
                 @Override
                 public void receive(Event[] events) {
+                    System.out.print("output event: ");
+                    EventPrinter.print(events);
 
+                    for (Event event : events) {
+                        //  assertArrayEquals((Object[]) results.get(count), event.getData());
+                        count++;
+                    }
+                }
+            });
+            InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+            executionPlanRuntime.start();
+            inputHandler.send(new Object[]{"IBM", 10f, 14});
+            inputHandler.send(new Object[]{"IBM", 20.5f, 2});
+            inputHandler.send(new Object[]{"WSO2", 20.5f, 1});
+            inputHandler.send(new Object[]{"WSO2", 23f, 1});
+
+            Thread.sleep(1000);
+        } finally {
+            executionPlanRuntime.shutdown();
+        }
+    }
+
+
+    @Test
+    public void testMaxByLengthWindowProcessor2() throws InterruptedException {
+        log.info("Testing maxByLengthWindowProcessor with no of events equal to window size for integer type parameter");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLength(volume, 4) select symbol,price," +
+                "volume insert into outputStream ;";
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+        try {
+            final List<Object> results = new ArrayList<Object>();
+            results.add(new Object[]{"IBM", 700f, 14});
+            results.add(new Object[]{"IBM", 700f, 14});
+            results.add(new Object[]{"IBM", 700f, 20});
+            results.add(new Object[]{"ZZZ", 60.5f, 82});
+            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+                @Override
+                public void receive(Event[] events) {
                     System.out.print("output event: ");
                     EventPrinter.print(events);
                     for (Event event : events) {
@@ -111,72 +158,71 @@ public class MaxByLengthBatchWindowProcessorTestCase {
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
             inputHandler.send(new Object[]{"IBM", 700f, 14});
-            inputHandler.send(new Object[]{"IBM", 60.5f, 2});
-            inputHandler.send(new Object[]{"IBM", 700f, 142});
-            inputHandler.send(new Object[]{"IBM", 60.5f, 21});
-            inputHandler.send(new Object[]{"et", 700f, 1});
-            inputHandler.send(new Object[]{"dg", 60.5f, 24});
-            inputHandler.send(new Object[]{"IBM", 60.5f, 21});
-            inputHandler.send(new Object[]{"et", 700f, 1});
-            inputHandler.send(new Object[]{"dg", 60.5f, 24});
+            inputHandler.send(new Object[]{"IBM", 60.5f, 12});
+            inputHandler.send(new Object[]{"IBM", 700f, 20});
+            inputHandler.send(new Object[]{"ZZZ", 60.5f, 82});
             Thread.sleep(1000);
-
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
+
     @Test
-    public void testMaxByWindowForLengthBatch3() throws InterruptedException {
-        log.info("Testing maxByLengthBatchWindowProcessor with no of events greater than window size for String parameter");
+    public void testMaxByLengthWindowProcessor3() throws InterruptedException {
+        log.info("Testing maxByLengthWindowProcessor with no of events greater than window size for String type parameter");
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxByLengthBatch(symbol, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:maxByLength(symbol, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         try {
             final List<Object> results = new ArrayList<Object>();
-            results.add(new Object[]{"XXX", 700f, 14});
-            results.add(new Object[]{"YTX", 60.5f, 24});
+            results.add(new Object[]{"bbc", 700f, 14});
+            results.add(new Object[]{"bbc", 700f, 14});
+            results.add(new Object[]{"xxx", 700f, 2});
+            results.add(new Object[]{"xxx", 700f, 2});
+            results.add(new Object[]{"xxx", 700f, 2});
+            results.add(new Object[]{"zzz", 60.5f, 12});
+            results.add(new Object[]{"zzz", 60.5f, 12});
+            results.add(new Object[]{"zzz", 60.5f, 12});
+            results.add(new Object[]{"zzz", 60.5f, 12});
+            results.add(new Object[]{"rye", 60.5f, 82});
 
             executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
-
                 @Override
                 public void receive(Event[] events) {
-
                     System.out.print("output event: ");
                     EventPrinter.print(events);
+
                     for (Event event : events) {
                         assertArrayEquals((Object[]) results.get(count), event.getData());
                         count++;
                     }
-
                 }
             });
             InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             executionPlanRuntime.start();
-            inputHandler.send(new Object[]{"XXX", 700f, 14});
-            inputHandler.send(new Object[]{"ABC", 60.5f, 2});
-            inputHandler.send(new Object[]{"AAA", 700f, 142});
-            inputHandler.send(new Object[]{"ACD", 60.5f, 21});
-            System.out.println("-------------------");
-            inputHandler.send(new Object[]{"RTE", 700f, 1});
-            inputHandler.send(new Object[]{"YTX", 60.5f, 24});
-            inputHandler.send(new Object[]{"DGF", 60.5f, 21});
-            inputHandler.send(new Object[]{"ETR", 700f, 1});
-            inputHandler.send(new Object[]{"DXD", 60.5f, 24});
-
+            inputHandler.send(new Object[]{"bbc", 700f, 14});
+            inputHandler.send(new Object[]{"bab", 60.5f, 12});
+            inputHandler.send(new Object[]{"xxx", 700f, 2});
+            inputHandler.send(new Object[]{"ddd", 60.5f, 82});
+            inputHandler.send(new Object[]{"abc", 700f, 84});
+            inputHandler.send(new Object[]{"zzz", 60.5f, 12});
+            inputHandler.send(new Object[]{"aaa", 700f, 2});
+            inputHandler.send(new Object[]{"dhh", 60.5f, 82});
+            inputHandler.send(new Object[]{"drg", 700f, 2});
+            inputHandler.send(new Object[]{"rye", 60.5f, 82});
             Thread.sleep(1000);
-
-
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
 
+
     @Test
-    public void testMaxByWindowForLengthBatch4() throws InterruptedException {
+    public void testMaxByWindowForLength4() throws InterruptedException {
         log.info("Join test1");
 
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -185,11 +231,10 @@ public class MaxByLengthBatchWindowProcessorTestCase {
                 "define stream twitterStream (user string, tweet string, company string,likes int); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.lengthBatch(3) join twitterStream#window.minbymaxby:maxByLengthBatch(likes, 2) " +
+                "from cseEventStream#window.lengthBatch(3) join twitterStream#window.extrema:maxByLength(likes, 2) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price, twitterStream.likes " +
                 "insert into outputStream ;";
-
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
         try {
             final List<Object> results = new ArrayList<Object>();
@@ -206,12 +251,12 @@ public class MaxByLengthBatchWindowProcessorTestCase {
             InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
             executionPlanRuntime.start();
 
-            cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
+            cseEventStreamHandler.send(new Object[]{"XXX", 699f, 14});
             cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
             cseEventStreamHandler.send(new Object[]{"WSO2", 700f, 142});
 
-            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2", 23});
-            twitterStreamHandler.send(new Object[]{"User1", "Hello SIDDHI", "WSO2", 56});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2", 43});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello SIDDHI", "WSO2", 45});
 
             cseEventStreamHandler.send(new Object[]{"ACD", 60.5f, 21});
             cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
@@ -223,7 +268,7 @@ public class MaxByLengthBatchWindowProcessorTestCase {
     }
 
     @Test
-    public void testMaxByWindowForLengthBatch6() throws InterruptedException {
+    public void testMaxByWindowForLength6() throws InterruptedException {
         log.info("Join test2");
 
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -232,10 +277,10 @@ public class MaxByLengthBatchWindowProcessorTestCase {
                 "define stream twitterStream (num float, tweet string, company string); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.minbymaxby:maxByLengthBatch(price, 2) join twitterStream#window.minbymaxby:maxByLengthBatch(num, 2) " +
+                "from cseEventStream#window.extrema:maxByLength(price, 2) join twitterStream#window.extrema:maxByLength(num, 2) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
-                "insert events into outputStream ;";
+                "insert all events into outputStream ;";
 
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
         try {
@@ -246,22 +291,24 @@ public class MaxByLengthBatchWindowProcessorTestCase {
                 public void receive(Event[] events) {
 
                     System.out.print("output event: ");
+//                    if(events==null){
+//                        System.out.println("There is no output events");
+//                    }
                     EventPrinter.print(events);
+
                 }
             });
             InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
             InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
             executionPlanRuntime.start();
 
-            cseEventStreamHandler.send(new Object[]{"WSO2", 700f, 14});
-            cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
+            cseEventStreamHandler.send(new Object[]{"WSO2", 10f, 14});
+            cseEventStreamHandler.send(new Object[]{"AAA", 60.5f, 2});
 
             twitterStreamHandler.send(new Object[]{100, "Hello World", "XXX"});
             twitterStreamHandler.send(new Object[]{101, "Hello SIDDHI", "WSO2"});
+            twitterStreamHandler.send(new Object[]{104, "Hello CEP", "WSO2"});
 
-            cseEventStreamHandler.send(new Object[]{"ACD", 60.5f, 21});
-            cseEventStreamHandler.send(new Object[]{"XXX", 700f, 14});
-            cseEventStreamHandler.send(new Object[]{"WSO2", 60.5f, 222});
 
         } finally {
             executionPlanRuntime.shutdown();
@@ -271,7 +318,7 @@ public class MaxByLengthBatchWindowProcessorTestCase {
 
 
     @Test
-    public void testMaxByWindowForLengthBatch7() throws InterruptedException {
+    public void testMaxByWindowForLength7() throws InterruptedException {
         log.info("Join test3");
 
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -280,21 +327,18 @@ public class MaxByLengthBatchWindowProcessorTestCase {
                 "define stream twitterStream (num float, tweet string, company string); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.minbymaxby:maxByLengthBatch(price, 2) join twitterStream#window.LengthBatch(2) " +
+                "from cseEventStream#window.extrema:maxByLength(price, 2) join twitterStream#window.Length(2) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
                 "insert events into outputStream ;";
 
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
         try {
-            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
 
+            executionPlanRuntime.addCallback("query1", new QueryCallback() {
                 @Override
-                public void receive(Event[] events) {
-
-                    System.out.print("output event: ");
-                    EventPrinter.print(events);
-
+                public void receive(long l, Event[] events, Event[] events1) {
+                    EventPrinter.print(l, events, events1);
                 }
             });
             InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
@@ -306,13 +350,19 @@ public class MaxByLengthBatchWindowProcessorTestCase {
 
             twitterStreamHandler.send(new Object[]{100, "Hello World", "XXX"});
             twitterStreamHandler.send(new Object[]{101, "Hello SIDDHI", "WSO2"});
+            Thread.sleep(100);
+            System.out.println(",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,");
 
             cseEventStreamHandler.send(new Object[]{"WSO2", 900f, 14});
-            cseEventStreamHandler.send(new Object[]{"ABC", 60.5f, 2});
+            System.out.println("------------------------------------");
+            cseEventStreamHandler.send(new Object[]{"XXX", 70.5f, 2});
+
 
         } finally {
             executionPlanRuntime.shutdown();
         }
 
     }
+
+
 }
