+++ src/test/java/org/wso2/extension/siddhi/execution/extrema/MinByLengthBatchWindowProcessorTestCase.java
@@ -1,6 +1,6 @@
 
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -15,8 +15,10 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
+ *
  */
-package org.wso2.extension.siddhi.window.minbymaxby.lengthwindow;
+
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.apache.log4j.Logger;
 import org.junit.Before;
@@ -54,7 +56,7 @@ public class MinByLengthBatchWindowProcessorTestCase {
 
         //siddhiManager.setExtension("unique:minByLengthBatch", MinByLengthBatchWindowProcessor.class);
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:minByLengthBatch(price, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minByLengthBatch(price, 4) select symbol,price," +
                 "volume insert into outputStream ;";
 
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
@@ -93,7 +95,7 @@ public class MinByLengthBatchWindowProcessorTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:minByLengthBatch(volume, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minByLengthBatch(volume, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         try {
@@ -138,7 +140,7 @@ public class MinByLengthBatchWindowProcessorTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:minByLengthBatch(symbol, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minByLengthBatch(symbol, 4) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         try {
@@ -181,7 +183,7 @@ public class MinByLengthBatchWindowProcessorTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:minByLengthBatch(price, 4) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minByLengthBatch(price, 4) select symbol,price," +
                 "volume insert into outputStream ;";
 
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
