+++ src/test/java/org/wso2/extension/siddhi/execution/extrema/MinByTimeWindowTestCase.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.extension.siddhi.window.minbymaxby.TimeWindow;
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -27,9 +27,10 @@ import org.wso2.siddhi.core.SiddhiManager;
 import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.query.output.callback.QueryCallback;
 import org.wso2.siddhi.core.stream.input.InputHandler;
+import org.wso2.siddhi.core.stream.output.StreamCallback;
 import org.wso2.siddhi.core.util.EventPrinter;
 
-public class MaxByTimeWindowTestCase {
+public class MinByTimeWindowTestCase {
 
     private int inEventCount;
     private int removeEventCount;
@@ -47,12 +48,12 @@ public class MaxByTimeWindowTestCase {
      */
 
     @Test
-    public void maxbyTimeWindowTest1() throws InterruptedException {
+    public void minbyTimeWindowTest1() throws InterruptedException {
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxbytime(price, 1 sec) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minbytime(price, 1 sec) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         executionPlanRuntime.addCallback("query1", new QueryCallback() {
@@ -63,7 +64,6 @@ public class MaxByTimeWindowTestCase {
                     inEventCount = inEventCount + inEvents.length;
                 }
                 if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
                     removeEventCount = removeEventCount + removeEvents.length;
                 }
                 eventArrived = true;
@@ -82,18 +82,18 @@ public class MaxByTimeWindowTestCase {
         inputHandler.send(new Object[]{"IBM", 60.50f, 6});
         inputHandler.send(new Object[]{"AAA", 600.5f, 7});
         Thread.sleep(500);
-        Assert.assertEquals(4, inEventCount);
+        Assert.assertEquals(5, inEventCount);
         Assert.assertEquals(0,removeEventCount);
         Assert.assertTrue(eventArrived);
         executionPlanRuntime.shutdown();
     }
 
     @Test
-    public void maxbyTimeWindowTest2() throws InterruptedException {
+    public void minbyTimeWindowTest2() throws InterruptedException {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxbytime(price, 1 sec) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minbytime(price, 1 sec) select symbol,price," +
                 "volume insert into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         executionPlanRuntime.addCallback("query1", new QueryCallback() {
@@ -104,7 +104,6 @@ public class MaxByTimeWindowTestCase {
                     inEventCount = inEventCount + inEvents.length;
                 }
                 if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
                     removeEventCount = removeEventCount + removeEvents.length;
                 }
                 eventArrived = true;
@@ -114,17 +113,17 @@ public class MaxByTimeWindowTestCase {
         InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
         executionPlanRuntime.start();
         inputHandler.send(new Object[]{"IBM", 700f, 1});
-        inputHandler.send(new Object[]{"MIT", 700f, 2});
-        inputHandler.send(new Object[]{"WSO2", 700f, 3});
+        inputHandler.send(new Object[]{"IBM", 798f, 1});
+        inputHandler.send(new Object[]{"IBM", 432f, 1});
         Thread.sleep(1100);
-        Assert.assertEquals(3, inEventCount);
+        Assert.assertEquals(2, inEventCount);
         Assert.assertEquals(0,removeEventCount);
         Assert.assertTrue(eventArrived);
         executionPlanRuntime.shutdown();
     }
 
     @Test
-    public void maxbyTimeWindowTest3() throws InterruptedException {
+    public void minbyTimeWindowTest3() throws InterruptedException {
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
@@ -132,9 +131,9 @@ public class MaxByTimeWindowTestCase {
                 "define stream cseEventStream (symbol string, price float, volume int);";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.minbymaxby:maxbytime(symbol, 1 sec) " +
+                "from cseEventStream#window.extrema:minbytime(price, 1 sec) " +
                 "select symbol, price " +
-                "insert into outputStream ;";
+                "insert expired events into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         executionPlanRuntime.addCallback("query1", new QueryCallback() {
             @Override
@@ -153,30 +152,35 @@ public class MaxByTimeWindowTestCase {
 
         InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
         executionPlanRuntime.start();
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
+        Thread.sleep(100);
         inputHandler.send(new Object[]{"WSO2", 60.5f, 2});
-        inputHandler.send(new Object[]{"MIT", 23.5f, 3});
+        Thread.sleep(100);
         inputHandler.send(new Object[]{"GOOGLE", 45.5f, 4});
-        Thread.sleep(1100);
+        Thread.sleep(100);
+        inputHandler.send(new Object[]{"MIT", 23.5f, 3});
+        Thread.sleep(2100);
+
         inputHandler.send(new Object[]{"ORACLE", 10f, 5});
         inputHandler.send(new Object[]{"WSO2", 34.5f, 6});
         inputHandler.send(new Object[]{"GOOGLE", 65.5f, 7});
         inputHandler.send(new Object[]{"MIT", 7.5f, 8});
-        Thread.sleep(1100);
+        Thread.sleep(2100);
         inputHandler.send(new Object[]{"GOOGLE", 7f, 9});
         inputHandler.send(new Object[]{"WSO2", 60.5f, 10});
         inputHandler.send(new Object[]{"MIT", 632.5f, 11});
         Thread.sleep(4000);
         executionPlanRuntime.shutdown();
         Assert.assertEquals(0,inEventCount);
-        Assert.assertEquals(3, removeEventCount);
+        Assert.assertEquals(7, removeEventCount);
 
     }
     @Test
-    public void maxbyTimeWindowTest4() throws InterruptedException {
+    public void minbyTimeWindowTest4() throws InterruptedException {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxbytime(price, 1 sec) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minbytime(price, 1 sec) select symbol,price," +
                 "volume insert expired events into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
         executionPlanRuntime.addCallback("query1", new QueryCallback() {
@@ -187,7 +191,6 @@ public class MaxByTimeWindowTestCase {
                     inEventCount = inEventCount + inEvents.length;
                 }
                 if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
                     removeEventCount = removeEventCount + removeEvents.length;
                 }
                 eventArrived = true;
@@ -200,7 +203,6 @@ public class MaxByTimeWindowTestCase {
         inputHandler.send(new Object[]{"WSO2", 798f, 2});
         inputHandler.send(new Object[]{"MIT", 432f, 3});
         Thread.sleep(1100);
-        executionPlanRuntime.start();
         inputHandler.send(new Object[]{"IFS", 700f, 4});
         inputHandler.send(new Object[]{"GOOGLE", 798f, 5});
         inputHandler.send(new Object[]{"YAHOO", 432f, 6});
@@ -208,14 +210,16 @@ public class MaxByTimeWindowTestCase {
         inputHandler.send(new Object[]{"YAHOO", 432f, 8});
         Thread.sleep(1100);
         executionPlanRuntime.shutdown();
+        Assert.assertEquals(0,inEventCount);
+        Assert.assertEquals(5, removeEventCount);
     }
 
     @Test
-    public void maxbyTimeWindowTest6() throws InterruptedException {
+    public void minbyTimeWindowTest6() throws InterruptedException {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.minbymaxby:maxbytime(price, 1 sec) select symbol,price," +
+        String query = "@info(name = 'query1') from cseEventStream#window.extrema:minbytime(price, 1 sec) select symbol,price," +
                 "volume insert all events into outputStream ;";
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
 
@@ -227,7 +231,6 @@ public class MaxByTimeWindowTestCase {
                     inEventCount = inEventCount + inEvents.length;
                 }
                 if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
                     removeEventCount = removeEventCount + removeEvents.length;
                 }
                 eventArrived = true;
@@ -248,11 +251,94 @@ public class MaxByTimeWindowTestCase {
         inputHandler.send(new Object[]{"YAHOO", 32f, 8});
         Thread.sleep(1100);
         executionPlanRuntime.shutdown();
-        Assert.assertEquals(4, inEventCount);
-        Assert.assertEquals(2, removeEventCount);
+        Assert.assertEquals(5, inEventCount);
+        Assert.assertEquals(5, removeEventCount);
+    }
+    @Test
+    public void minbyTimeWindowTest5() throws InterruptedException {
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String streams = "" +
+                "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define stream twitterStream (user string, tweet string, company string); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from twitterStream#window.extrema:minbytime(company,1 sec) join cseEventStream#window.length(2) " +
+                "on cseEventStream.symbol == twitterStream.company " +
+                "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
+                "insert into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
+        try {
+
+            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+                @Override
+                public void receive(Event[] events) {
+                    EventPrinter.print(events);
+                }
+            });
+            InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+            InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
+            executionPlanRuntime.start();
+            cseEventStreamHandler.send(new Object[]{"IBM", 75.6f, 100});
+            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
+
+            twitterStreamHandler.send(new Object[]{"User2", "Hi", "IBM"});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+
+            Thread.sleep(1500);
+//            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+//            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
+//            Thread.sleep(1000);
+            //Assert.assertTrue("In Events can be 1 or 2 ", inEventCount == 1 || inEventCount == 2);
+            Assert.assertEquals(0, removeEventCount);
+//            Assert.assertTrue(eventArrived);
+        } finally {
+            executionPlanRuntime.shutdown();
+        }
     }
 
+    @Test
+    public void minbyTimeWindowTest7() throws InterruptedException {
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String streams = "" +
+                "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define stream twitterStream (user string, tweet string, company string); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from twitterStream#window.extrema:maxbytime(company,1 sec) join cseEventStream#window.extrema:minbytime(price,1 sec) " +
+                "on cseEventStream.symbol == twitterStream.company " +
+                "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
+                "insert into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
+        try {
+
+            executionPlanRuntime.addCallback("outputStream", new StreamCallback() {
+                @Override
+                public void receive(Event[] events) {
+                    EventPrinter.print(events);
+                }
+            });
+            InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+            InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
+            executionPlanRuntime.start();
+            cseEventStreamHandler.send(new Object[]{"IBM", 25.0f, 100});
+            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
+            twitterStreamHandler.send(new Object[]{"User2", "Hi", "IBM"});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+            Thread.sleep(1500);
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
+            Thread.sleep(1000);
+            //Assert.assertTrue("In Events can be 1 or 2 ", inEventCount == 1 || inEventCount == 2);
+            Assert.assertEquals(0, removeEventCount);
+//            Assert.assertTrue(eventArrived);
+        } finally {
+            executionPlanRuntime.shutdown();
+        }
+    }
 
 }
+
